requirements.txt
flask
google-genai==1.2.0
requests
gunicorn
